+++ src/main/java/de/markusbordihn/lobby/commands/LobbyCommand.java
@@ -45,9 +45,6 @@ import de.markusbordihn.lobby.teleporter.PlayerTeleportManager;
 @EventBusSubscriber
 public class LobbyCommand extends CustomCommand {
 
-  public static final String DIMENSION_NAME = "Lobby";
-  public static final int PERMISSION_LEVEL = 0;
-
   private static final CommonConfig.Config COMMON = CommonConfig.COMMON;
   private static boolean lobbyRestrictCommand = COMMON.lobbyRestrictCommand.get();
   private static boolean teleportDelayEnabled = COMMON.teleportDelayEnabled.get();
@@ -70,9 +67,11 @@ public class LobbyCommand extends CustomCommand {
     if (Boolean.FALSE.equals(COMMON.lobbyEnabled.get())) {
       return;
     }
-    registerCommand(COMMON.lobbyCommandName.get(), DIMENSION_NAME, PERMISSION_LEVEL);
+    registerCommand(COMMON.lobbyCommandName.get(), COMMON.lobbyDimensionName.get(),
+        COMMON.lobbyCommandPermissionLevel.get());
     dispatcher.register(Commands.literal(COMMON.lobbyCommandName.get())
-        .requires(cs -> cs.hasPermission(PERMISSION_LEVEL)).executes(command));
+        .requires(cs -> cs.hasPermission(COMMON.lobbyCommandPermissionLevel.get()))
+        .executes(command));
   }
 
   @Override
@@ -83,8 +82,10 @@ public class LobbyCommand extends CustomCommand {
     Long coolDownTimer = coolDownPlayerMap.getOrDefault(player, null);
     Long currentTimer = java.time.Instant.now().getEpochSecond();
     if (coolDownTimer != null && coolDownTimer > currentTimer) {
-      sendFeedback(context, new TranslatableComponent(Constants.TELEPORT_FAILED_COOLDOWN,
-          DIMENSION_NAME, coolDownTimer - currentTimer).withStyle(ChatFormatting.RED));
+      sendFeedback(context,
+          new TranslatableComponent(Constants.TELEPORT_FAILED_COOLDOWN,
+              COMMON.lobbyDimensionName.get(), coolDownTimer - currentTimer)
+                  .withStyle(ChatFormatting.RED));
       return 0;
     } else {
       coolDownPlayerMap.put(player, currentTimer + generalCommandCoolDown);
@@ -92,23 +93,26 @@ public class LobbyCommand extends CustomCommand {
 
     // Provide feedback to the player for their teleporter request.
     if (DimensionManager.getLobbyDimension() == null) {
-      sendFeedback(context, new TranslatableComponent(Constants.UNABLE_TO_TELEPORT_MESSAGE,
-          DIMENSION_NAME, DimensionManager.getLobbyDimensionName()).withStyle(ChatFormatting.RED));
+      sendFeedback(context,
+          new TranslatableComponent(Constants.UNABLE_TO_TELEPORT_MESSAGE,
+              COMMON.lobbyDimensionName.get(), DimensionManager.getLobbyDimensionName())
+                  .withStyle(ChatFormatting.RED));
     } else if (!lobbyRestrictCommand || player.getLevel() != DimensionManager.getLobbyDimension()) {
       if (teleportDelayEnabled && teleportDelayCounter > 0) {
-        sendFeedback(context, new TranslatableComponent(Constants.TELEPORT_TO_IN_MESSAGE,
-            DIMENSION_NAME, teleportDelayCounter).withStyle(ChatFormatting.GREEN));
+        sendFeedback(context,
+            new TranslatableComponent(Constants.TELEPORT_TO_IN_MESSAGE,
+                COMMON.lobbyDimensionName.get(), teleportDelayCounter)
+                    .withStyle(ChatFormatting.GREEN));
         PlayerTeleportManager.teleportPlayerToLobby(player);
       } else {
-        sendFeedback(context,
-            new TranslatableComponent(Constants.TELEPORT_TO_MESSAGE, DIMENSION_NAME)
-                .withStyle(ChatFormatting.GREEN));
+        sendFeedback(context, new TranslatableComponent(Constants.TELEPORT_TO_MESSAGE,
+            COMMON.lobbyDimensionName.get()).withStyle(ChatFormatting.GREEN));
         DimensionManager.teleportToLobby(player);
       }
     } else {
       sendFeedback(context,
           new TranslatableComponent(Constants.TELEPORT_FAILED_ALREADY_IN_DIMENSION_MESSAGE,
-              DIMENSION_NAME).withStyle(ChatFormatting.YELLOW));
+              COMMON.lobbyDimensionName.get()).withStyle(ChatFormatting.YELLOW));
     }
     return 0;
   }