+++ src/main/java/de/markusbordihn/lobby/commands/VoidCommand.java
@@ -45,9 +45,6 @@ import de.markusbordihn.lobby.teleporter.PlayerTeleportManager;
 @EventBusSubscriber
 public class VoidCommand extends CustomCommand {
 
-  public static final String DIMENSION_NAME = "Void";
-  public static final int PERMISSION_LEVEL = 0;
-
   private static final CommonConfig.Config COMMON = CommonConfig.COMMON;
   private static boolean voidRestrictCommand = COMMON.voidRestrictCommand.get();
   private static boolean teleportDelayEnabled = COMMON.teleportDelayEnabled.get();
@@ -70,9 +67,11 @@ public class VoidCommand extends CustomCommand {
     if (Boolean.FALSE.equals(COMMON.voidEnabled.get())) {
       return;
     }
-    registerCommand(COMMON.voidCommandName.get(), DIMENSION_NAME, PERMISSION_LEVEL);
+    registerCommand(COMMON.voidCommandName.get(), COMMON.voidDimensionName.get(),
+        COMMON.voidCommandPermissionLevel.get());
     dispatcher.register(Commands.literal(COMMON.voidCommandName.get())
-        .requires(cs -> cs.hasPermission(PERMISSION_LEVEL)).executes(command));
+        .requires(cs -> cs.hasPermission(COMMON.voidCommandPermissionLevel.get()))
+        .executes(command));
   }
 
   @Override
@@ -83,8 +82,10 @@ public class VoidCommand extends CustomCommand {
     Long coolDownTimer = coolDownPlayerMap.getOrDefault(player, null);
     Long currentTimer = java.time.Instant.now().getEpochSecond();
     if (coolDownTimer != null && coolDownTimer > currentTimer) {
-      sendFeedback(context, new TranslatableComponent(Constants.TELEPORT_FAILED_COOLDOWN,
-          DIMENSION_NAME, coolDownTimer - currentTimer).withStyle(ChatFormatting.RED));
+      sendFeedback(context,
+          new TranslatableComponent(Constants.TELEPORT_FAILED_COOLDOWN,
+              COMMON.voidDimensionName.get(), coolDownTimer - currentTimer)
+                  .withStyle(ChatFormatting.RED));
       return 0;
     } else {
       coolDownPlayerMap.put(player, currentTimer + generalCommandCoolDown);
@@ -93,22 +94,24 @@ public class VoidCommand extends CustomCommand {
     // Provide feedback to the player for their teleporter request.
     if (DimensionManager.getVoidDimension() == null) {
       sendFeedback(context, new TranslatableComponent(Constants.UNABLE_TO_TELEPORT_MESSAGE,
-          DIMENSION_NAME, DimensionManager.getVoidDimensionName()));
+          COMMON.voidDimensionName.get(), DimensionManager.getVoidDimensionName()));
     } else if (!voidRestrictCommand || player.getLevel() != DimensionManager.getVoidDimension()) {
       if (teleportDelayEnabled && teleportDelayCounter > 0) {
-        sendFeedback(context, new TranslatableComponent(Constants.TELEPORT_TO_IN_MESSAGE,
-            DIMENSION_NAME, teleportDelayCounter).withStyle(ChatFormatting.GREEN));
+        sendFeedback(context,
+            new TranslatableComponent(Constants.TELEPORT_TO_IN_MESSAGE,
+                COMMON.voidDimensionName.get(), teleportDelayCounter)
+                    .withStyle(ChatFormatting.GREEN));
         PlayerTeleportManager.teleportPlayerToVoid(player);
       } else {
         sendFeedback(context,
-            new TranslatableComponent(Constants.TELEPORT_TO_MESSAGE, DIMENSION_NAME)
+            new TranslatableComponent(Constants.TELEPORT_TO_MESSAGE, COMMON.voidDimensionName.get())
                 .withStyle(ChatFormatting.GREEN));
         DimensionManager.teleportToVoid(player);
       }
     } else {
       sendFeedback(context,
           new TranslatableComponent(Constants.TELEPORT_FAILED_ALREADY_IN_DIMENSION_MESSAGE,
-              DIMENSION_NAME).withStyle(ChatFormatting.YELLOW));
+              COMMON.voidDimensionName.get()).withStyle(ChatFormatting.YELLOW));
     }
     return 0;
   }