+++ src/main/java/de/markusbordihn/lobby/dimension/DimensionManager.java
@@ -26,33 +26,17 @@ import java.util.concurrent.ConcurrentHashMap;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import net.minecraft.core.BlockPos;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.ExperienceOrb;
-import net.minecraft.world.entity.LightningBolt;
-import net.minecraft.world.entity.ambient.Bat;
-import net.minecraft.world.entity.item.FallingBlockEntity;
-import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.entity.projectile.Projectile;
-import net.minecraft.world.entity.vehicle.MinecartChest;
-import net.minecraft.world.level.BaseSpawner;
 import net.minecraft.world.level.GameType;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelAccessor;
 
-import net.minecraftforge.event.entity.EntityJoinWorldEvent;
-import net.minecraftforge.event.entity.living.LivingSpawnEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent.PlayerChangedDimensionEvent;
 import net.minecraftforge.event.server.ServerAboutToStartEvent;
 import net.minecraftforge.event.server.ServerStartedEvent;
-import net.minecraftforge.eventbus.api.Event;
-import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.server.ServerLifecycleHooks;
@@ -76,30 +60,21 @@ public class DimensionManager {
 
   private static boolean fishingEnabled = COMMON.fishingEnabled.get();
   private static String fishingDimension = COMMON.fishingDimension.get();
-  private static boolean fishingDisableMobSpawning = COMMON.fishingDisableMobSpawning.get();
   private static List<String> fishingBuilderList = COMMON.fishingBuilderList.get();
 
   private static boolean gamingEnabled = COMMON.gamingEnabled.get();
   private static String gamingDimension = COMMON.gamingDimension.get();
-  private static boolean gamingDisableMobSpawning = COMMON.gamingDisableMobSpawning.get();
   private static List<String> gamingBuilderList = COMMON.gamingBuilderList.get();
 
   private static boolean lobbyEnabled = COMMON.lobbyEnabled.get();
   private static String lobbyDimension = COMMON.lobbyDimension.get();
-  private static boolean lobbyDisableMobSpawning = COMMON.lobbyDisableMobSpawning.get();
   private static List<String> lobbyBuilderList = COMMON.lobbyBuilderList.get();
 
   private static boolean miningEnabled = COMMON.miningEnabled.get();
   private static String miningDimension = COMMON.miningDimension.get();
-  private static boolean miningDisableBatSpawning = COMMON.miningDisableBatSpawning.get();
-  private static boolean miningDisableMobSpawning = COMMON.miningDisableMobSpawning.get();
-  private static boolean miningDisableMinecartChestSpawning =
-      COMMON.miningDisableMinecartChestSpawning.get();
-  private static boolean miningRemoveSpawner = COMMON.miningRemoveSpawner.get();
 
   private static boolean voidEnabled = COMMON.voidEnabled.get();
   private static String voidDimension = COMMON.voidDimension.get();
-  private static boolean voidDisableMobSpawning = COMMON.voidDisableMobSpawning.get();
   private static List<String> voidBuilderList = COMMON.voidBuilderList.get();
 
   private static Set<ServerPlayer> gameTypeReset = ConcurrentHashMap.newKeySet();
@@ -130,29 +105,21 @@ public class DimensionManager {
 
     fishingEnabled = COMMON.fishingEnabled.get();
     fishingDimension = COMMON.fishingDimension.get();
-    fishingDisableMobSpawning = COMMON.fishingDisableMobSpawning.get();
     fishingBuilderList = COMMON.fishingBuilderList.get();
 
     gamingEnabled = COMMON.gamingEnabled.get();
     gamingDimension = COMMON.gamingDimension.get();
-    gamingDisableMobSpawning = COMMON.gamingDisableMobSpawning.get();
     gamingBuilderList = COMMON.gamingBuilderList.get();
 
     lobbyEnabled = COMMON.lobbyEnabled.get();
     lobbyDimension = COMMON.lobbyDimension.get();
-    lobbyDisableMobSpawning = COMMON.lobbyDisableMobSpawning.get();
     lobbyBuilderList = COMMON.lobbyBuilderList.get();
 
     miningEnabled = COMMON.miningEnabled.get();
     miningDimension = COMMON.miningDimension.get();
-    miningDisableBatSpawning = COMMON.miningDisableBatSpawning.get();
-    miningDisableMobSpawning = COMMON.miningDisableMobSpawning.get();
-    miningDisableMinecartChestSpawning = COMMON.miningDisableMinecartChestSpawning.get();
-    miningRemoveSpawner = COMMON.miningRemoveSpawner.get();
 
     voidEnabled = COMMON.voidEnabled.get();
     voidDimension = COMMON.voidDimension.get();
-    voidDisableMobSpawning = COMMON.voidDisableMobSpawning.get();
     voidBuilderList = COMMON.voidBuilderList.get();
   }
 
@@ -180,27 +147,6 @@ public class DimensionManager {
     if (defaultHeal > 0) {
       log.info("{} Enable heal for default dimension for {} ticks.",
           Constants.LOG_DIMENSION_MANAGER_PREFIX, defaultHeal);
-    }
-
-    if (fishingDisableMobSpawning) {
-      log.info("{} Disable Mob Spawning for fishing dimension.",
-          Constants.LOG_DIMENSION_MANAGER_PREFIX);
-    }
-    if (lobbyDisableMobSpawning) {
-      log.info("{} Disable mob spawning for lobby dimension.",
-          Constants.LOG_DIMENSION_MANAGER_PREFIX);
-    }
-    if (miningDisableBatSpawning) {
-      log.info("{} Disable bat spawning for mining dimension.",
-          Constants.LOG_DIMENSION_MANAGER_PREFIX);
-    }
-    if (miningDisableMinecartChestSpawning) {
-      log.info("{} Disable minecraft chest spawning for mining dimension.",
-          Constants.LOG_DIMENSION_MANAGER_PREFIX);
-    }
-    if (miningDisableMobSpawning) {
-      log.info("{} Disable mob spawning for mining dimension.",
-          Constants.LOG_DIMENSION_MANAGER_PREFIX);
     }
   }
 
@@ -299,45 +245,6 @@ public class DimensionManager {
         gameTypeReset.remove(serverPlayer);
       }
     }
-  }
-
-  @SubscribeEvent(priority = EventPriority.HIGHEST)
-  public static void handleEntityJoinWorldEvent(EntityJoinWorldEvent event) {
-
-    // Ignore client side and everything which is not the mining dimension.
-    Level level = event.getWorld();
-    String dimensionLocation = level.dimension().location().toString();
-    if (level.isClientSide() || !dimensionLocation.equals(miningDimension)) {
-      return;
-    }
-
-    // Ignore specific entities and deny spawn of all others.
-    Entity entity = event.getEntity();
-    if (entity instanceof ItemEntity || entity instanceof ExperienceOrb
-        || entity instanceof LightningBolt || entity instanceof FallingBlockEntity
-        || entity instanceof Projectile || entity instanceof Player) {
-      return;
-    }
-
-    // Allow/deny Minecart Chest spawning
-    if (miningDisableMinecartChestSpawning && entity instanceof MinecartChest) {
-      event.setResult(Event.Result.DENY);
-    }
-
-    // Allow/deny Mob spawning
-    if (miningDisableMobSpawning) {
-      event.setResult(Event.Result.DENY);
-    }
-  }
-
-  @SubscribeEvent(priority = EventPriority.HIGHEST)
-  public static void handleLivingCheckSpawnEvent(LivingSpawnEvent.CheckSpawn event) {
-    handleSpawnEvent(event);
-  }
-
-  @SubscribeEvent(priority = EventPriority.HIGHEST)
-  public static void handleLivingSpecialSpawnEvent(LivingSpawnEvent.SpecialSpawn event) {
-    handleSpawnEvent(event);
   }
 
   private static void mapServerLevel(MinecraftServer server) {
@@ -474,6 +381,10 @@ public class DimensionManager {
     return miningLevel;
   }
 
+  public static ServerLevel getMiningDimensionRaw() {
+    return miningLevel;
+  }
+
   public static String getMiningDimensionName() {
     return miningDimension;
   }
@@ -584,95 +495,6 @@ public class DimensionManager {
 
   private static boolean shouldChangeGameType(Player player) {
     return !(player.isSpectator() || (player.hasPermissions(2) && player.isCreative()));
-  }
-
-  private static void handleSpawnEvent(LivingSpawnEvent event) {
-
-    // Ignore client side.
-    LevelAccessor level = event.getWorld();
-    if (level.isClientSide()) {
-      return;
-    }
-
-    // Ignore null entities and specific entities.
-    Entity entity = event.getEntity();
-    if (entity == null || entity instanceof Projectile) {
-      return;
-    }
-
-    // Control spawns depending on the dimension.
-    String dimensionLocation = entity.getLevel().dimension().location().toString();
-    if (fishingDisableMobSpawning && fishingDimension.equals(dimensionLocation)) {
-      handleSpawnEventFishing(event);
-    } else if (lobbyDisableMobSpawning && lobbyDimension.equals(dimensionLocation)) {
-      handleSpawnEventLobby(event);
-    } else if (gamingDisableMobSpawning && gamingDimension.equals(dimensionLocation)) {
-      handleSpawnEventGaming(event);
-    } else if ((miningDisableBatSpawning || miningDisableMinecartChestSpawning
-        || miningDisableMobSpawning) && miningDimension.equals(dimensionLocation)) {
-      handleSpawnEventMining(level, entity, event);
-    } else if (voidDisableMobSpawning && voidDimension.equals(dimensionLocation)) {
-      handleSpawnEventVoid(event);
-    }
-  }
-
-  private static void handleSpawnEventFishing(LivingSpawnEvent event) {
-    if (fishingDisableMobSpawning) {
-      event.setResult(Event.Result.DENY);
-    }
-  }
-
-  private static void handleSpawnEventGaming(LivingSpawnEvent event) {
-    if (gamingDisableMobSpawning) {
-      event.setResult(Event.Result.DENY);
-    }
-  }
-
-  private static void handleSpawnEventLobby(LivingSpawnEvent event) {
-    if (lobbyDisableMobSpawning) {
-      event.setResult(Event.Result.DENY);
-    }
-  }
-
-  private static void handleSpawnEventMining(LevelAccessor level, Entity entity,
-      LivingSpawnEvent event) {
-    // Removing spawners as soon they try to spawn something.
-    if (miningRemoveSpawner && event instanceof LivingSpawnEvent.CheckSpawn checkSpawn
-        && checkSpawn.getSpawner() != null) {
-      BaseSpawner spawner = checkSpawn.getSpawner();
-      BlockPos blockPos = spawner.getSpawnerBlockEntity().getBlockPos();
-      if (blockPos != null) {
-        log.debug("{} Removing spawner {} at {}", Constants.LOG_DIMENSION_MANAGER_PREFIX, spawner,
-            blockPos);
-        level.removeBlock(blockPos, true);
-      }
-    }
-
-    // Allow/deny bat spawning for better cave experience
-    if (entity instanceof Bat) {
-      if (miningDisableBatSpawning) {
-        event.setResult(Event.Result.DENY);
-      }
-      return;
-    }
-
-    // Allow/deny Minecart Chest spawning
-    if (entity instanceof MinecartChest) {
-      if (miningDisableMinecartChestSpawning) {
-        event.setResult(Event.Result.DENY);
-      }
-      return;
-    }
-
-    if (miningDisableMobSpawning) {
-      event.setResult(Event.Result.DENY);
-    }
-  }
-
-  private static void handleSpawnEventVoid(LivingSpawnEvent event) {
-    if (voidDisableMobSpawning) {
-      event.setResult(Event.Result.DENY);
-    }
   }
 
 }